# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PC 견적 추천 API",
    "version": "2.0.0",
    "description": "PC 부품 조회·검색·비교, 가격 추세·멀티몰 비교, 견적 저장/공유/PDF, 가격 알림, 호환성 검사, AI 추천·업그레이드 어드바이저 API"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "components": {
...
    "parameters": {
      "category": {
        "name": "category", "in": "path", "required": true,
        "schema": { "type": "string", "enum": ["cpu","gpu","motherboard","memory","storage","psu","case","cooler"] }
      },
      "partName": {
        "name": "name", "in": "path", "required": true,
        "schema": { "type": "string" },
        "description": "부품 이름 (URL 인코딩)"
      }
    }
  },
  "paths": {
...
    "/api/parts": {
      "get": {
        "tags": ["parts"],
        "summary": "부품 목록 조회 (페이징)",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["cpu","gpu","motherboard","memory","storage","psu","case","cooler"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100, "maximum": 200 } }
        ],
        "responses": {
          "200": { "description": "부품 목록 (헤더: X-Total-Count, X-Page, X-Total-Pages)" }
        }
      }
    },
    "/api/parts/search": {
      "get": {
        "tags": ["parts"],
        "summary": "부품 검색 (이름·카테고리·가격범위·정렬)",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "검색어 (이름 포함 매칭)" },
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["cpu","gpu","motherboard","memory","storage","psu","case","cooler"] } },
          { "name": "manufacturer", "in": "query", "schema": { "type": "string" }, "description": "제조사 필터" },
          { "name": "priceMin", "in": "query", "schema": { "type": "integer" } },
          { "name": "priceMax", "in": "query", "schema": { "type": "integer" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["price_asc","price_desc","score_desc","value_desc"], "default": "price_asc" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "검색 결과 배열" },
          "400": { "description": "q 또는 category 중 하나 필수" }
        }
      }
    },
    "/api/parts/value-rank": {
      "get": {
        "tags": ["parts"],
        "summary": "가성비 랭킹 (벤치마크 점수 ÷ 가격)",
        "parameters": [
          { "name": "category", "in": "query", "required": true, "schema": { "type": "string", "enum": ["cpu","gpu","memory","storage"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 50 } }
        ],
        "responses": {
          "200": { "description": "가성비 순 부품 목록 (_valueScore 필드 포함)" }
        }
      }
    },
    "/api/parts/budget-picks": {
      "get": {
        "tags": ["parts"],
        "summary": "예산별 카테고리별 최적 부품 추천 (MongoDB $facet 집계)",
        "parameters": [
          { "name": "budget", "in": "query", "required": true, "schema": { "type": "integer", "example": 1500000 }, "description": "총 예산 원화" }
        ],
        "responses": {
          "200": {
            "description": "예산 배분 및 카테고리별 TOP 3",
            "content": {
              "application/json": {
                "example": {
                  "budget": 1500000,
                  "budgetAllocation": { "cpu": 375000, "gpu": 600000 },
                  "picks": { "cpu": [], "gpu": [] }
                }
              }
            }
          }
        }
      }
    },
    "/api/parts/batch": {
      "post": {
        "tags": ["parts"],
        "summary": "부품 일괄 조회 (최대 50개)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "items": [
                  { "category": "cpu", "name": "AMD Ryzen 5 7600X" },
                  { "category": "gpu", "name": "ASUS TUF GAMING RTX 4070" }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "조회된 부품 배열 (찾지 못한 항목 제외)" }
        }
      }
    },
    "/api/parts/compare": {
      "post": {
        "tags": ["parts"],
        "summary": "부품 비교 (2~5개, 가격·점수·가성비 비교)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "parts": [
                  { "category": "cpu", "name": "AMD Ryzen 5 7600X" },
                  { "category": "cpu", "name": "Intel Core i5-13600K" }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "비교 결과 (_isBestPrice, _isBestScore, _isBestValue, _scoreVsMax 포함)",
            "content": {
              "application/json": {
                "example": {
                  "count": 2,
                  "parts": [
                    { "name": "AMD Ryzen 5 7600X", "price": 250000, "_totalScore": 25000, "_valueScore": 0.1, "_isBestValue": true }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/parts/danawa-url": {
      "get": {
        "tags": ["parts"],
        "summary": "다나와 검색 URL 생성",
        "parameters": [
          { "name": "name", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "{ url, name }" }
        }
      }
    },
    "/api/parts/{category}/{name}": {
      "get": {
        "tags": ["parts"],
        "summary": "부품 상세 조회",
        "parameters": [
          { "$ref": "#/components/parameters/category" },
          { "$ref": "#/components/parameters/partName" }
        ],
        "responses": {
          "200": { "description": "부품 상세 정보" },
          "404": { "description": "부품 없음" }
        }
      }
    },
    "/api/parts/{category}/{name}/history": {
      "get": {
        "tags": ["parts"],
        "summary": "부품 가격 히스토리 전체 조회",
        "parameters": [
          { "$ref": "#/components/parameters/category" },
          { "$ref": "#/components/parameters/partName" }
        ],
        "responses": {
          "200": { "description": "{ name, category, priceHistory[] }" }
        }
      }
    },
    "/api/parts/{category}/{name}/trend": {
      "get": {
        "tags": ["parts"],
        "summary": "가격 추세 분석 (30·60·90일 최저·최고·평균·변동률)",
        "parameters": [
          { "$ref": "#/components/parameters/category" },
          { "$ref": "#/components/parameters/partName" }
        ],
        "responses": {
          "200": {
            "description": "기간별 가격 통계",
            "content": {
              "application/json": {
                "example": {
                  "name": "AMD Ryzen 5 7600X",
                  "category": "cpu",
                  "currentPrice": 250000,
                  "trends": [
                    { "days": 30, "min": 240000, "max": 260000, "avg": 250000, "change": -2.5, "count": 30 }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/prices/{category}/{name}": {
      "get": {
        "tags": ["prices"],
        "summary": "멀티몰 실시간 가격 비교 (네이버 쇼핑, 1시간 캐시)",
        "parameters": [
          { "$ref": "#/components/parameters/category" },
          { "$ref": "#/components/parameters/partName" }
        ],
        "responses": {
          "200": {
            "description": "멀티몰 가격 목록 + 최저가 요약",
            "content": {
              "application/json": {
                "example": {
                  "name": "AMD Ryzen 5 7600X",
                  "category": "cpu",
                  "storedPrice": 250000,
                  "lowestPrice": 243000,
                  "lowestMall": "쿠팡",
                  "priceGap": 7000,
                  "inStock": true,
                  "mallCount": 8,
                  "naverMalls": [
                    { "mallName": "쿠팡", "price": 243000, "link": "https://..." }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/prices/batch": {
      "post": {
        "tags": ["prices"],
        "summary": "빌드 전체 멀티몰 가격 일괄 조회 (최대 8개)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "parts": [
                  { "category": "cpu", "name": "AMD Ryzen 5 7600X" },
                  { "category": "gpu", "name": "ASUS TUF GAMING RTX 4070" }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "부품별 최저가 + 합계 요약",
            "content": {
              "application/json": {
                "example": {
                  "parts": [],
                  "summary": {
                    "totalLowest": 1200000,
                    "totalStored": 1250000,
                    "saving": 50000,
                    "allInStock": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}